microgridsToolkit/requirements.txt
Django==3.2.14
django_crispy_forms==1.14.0
hashids==1.3.1
numpy_financial==1.0.0
